Moves reviews.vue file
